+++ src/com/github/mjkuranda/retrographicsmanager/palettegenerator/PaletteWriter.java
@@ -1,5 +1,7 @@
 package com.github.mjkuranda.retrographicsmanager.palettegenerator;
 
+import com.github.mjkuranda.retrographicsmanager.palettes.Palette;
+
 import javax.imageio.ImageIO;
 import java.io.File;
 import java.io.IOException;