+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2023-05-10
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Abstraction for Palette objects.
+
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): RetroGraphicsManager handles all actions instead of Main class.
+
+### Removed
+- [@mjkuranda](https://github.com/mjkuranda): PaletteGenerator.
+
 ## [2.1.0] - 2023-05-08
 ### Changed
 - [@mjkuranda](https://github.com/mjkuranda): New color converting system.